Tutorials/requirements.txt
torch
torchvision
torchaudio
pybind11
opencv-python
matplotlib
Flask
blinker==1.4 
paho-mqtt
pygobject==3.50.0
pytest
pytest-cov
